mds/requirements.txt
fastapi
uvicorn[standard]
httpx
passlib
python-jose
python-multipart
websockets